# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentación realizado con Swagger, servicios funcionando con Laravel 8",
        "description": "Proyecto para la comunicación REST del sistema de Inventarios",
        "contact": {
            "name": "Correo de contacto",
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8001/SISHRS/backend/public"
        }
    ],
    "paths": {
...
        "/service/rest/v1/catalogo/typeroom": {
            "get": {
                "tags": [
                    "Type Room"
                ],
                "summary": "Muestra todos los tipos de habitación registradas en la base de datos.",
                "description": "Retorna un array de tipos de habitación.",
                "operationId": "getTypeRooms",
                "responses": {
                    "200": {
                        "description": "Respuesta correcta",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Permisos denegados"
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Type Room"
                ],
                "summary": "Crear un nuevo tipo de habitación en el sistema.",
                "description": "Retorna el objeto del tipo de habitación creado.",
                "operationId": "postTypeRoom",
                "responses": {
                    "200": {
                        "description": "Respuesta correcta",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Permisos denegados"
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/service/rest/v1/catalogo/typeroom/{typeroom}": {
            "put": {
                "tags": [
                    "Type Room"
                ],
                "summary": "Actualizar el tipo de habitación seleccionado.",
                "description": "Retorna el objeto del tipo de habitación actualizado.",
                "operationId": "updateTypeRoom",
                "parameters": [
                    {
                        "name": "typeroom",
                        "in": "path",
                        "description": "ID del tipo de habitación para actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Respuesta correcta",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Permisos denegados"
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/service/rest/v1/catalogo/typeroom/{typeroom}   ": {
            "delete": {
                "tags": [
                    "Type Room"
                ],
                "summary": "Eliminar el tipo de habitación seleccionado.",
                "description": "Retorna el objeto del tipo de habitación eliminado.",
                "operationId": "deleteTypeRoom",
                "parameters": [
                    {
                        "name": "typeroom",
                        "in": "path",
                        "description": "ID del tipo de habitación para eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Respuesta correcta",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Permisos denegados"
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "security": [
        []
    ]
}